+++ src/countKeyOccurrences.ts
@@ -1,0 +1,28 @@
+/**
+ * The function counts the number of occurrences of a specific key in a JSON object or array.
+ * @param {any} json - The `json` parameter is the JSON object or array that you want to search for the
+ * specified key in.
+ * @param {string} key - The `key` parameter is a string that represents the key you want to count in
+ * the JSON object.
+ * @returns the count of how many times the specified key appears in the given JSON object.
+ */
+export default function countKeyOccurrences(json: any, key: string) {
+  let count = 0;
+
+  if (typeof json === 'object' && !Array.isArray(json)) {
+    const keys = Object.keys(json);
+    if (keys.includes(key)) count += 1;
+
+    keys.forEach(k => {
+      count += countKeyOccurrences(json[k], key);
+    });
+  }
+
+  if (Array.isArray(json)) {
+    json.forEach(item => {
+      count += countKeyOccurrences(item, key);
+    });
+  }
+
+  return count;
+}
