+++ src/getFileExtension.ts
@@ -1,0 +1,16 @@
+/**
+ * The function `getFileExtension` takes a URL as input and returns the file extension of the URL, or
+ * an empty string if no extension is found.
+ * @param {string | undefined} url - The `url` parameter is a string that represents the URL of a file.
+ * @returns the file extension of a given URL as a lowercase string. If the URL is undefined or empty,
+ * an empty string is returned. If no file extension is found in the URL, null is returned.
+ */
+export default function getFileExtension(url: string | undefined): string | null {
+  if (!url) return '';
+  const lastDotIndex = url.lastIndexOf('.');
+  if (lastDotIndex === -1) {
+    return null; // No file extension found
+  }
+  const extension = url.slice(lastDotIndex + 1);
+  return extension.toLowerCase();
+}
