+++ src/isJSONObject.ts
@@ -1,0 +1,17 @@
+/**
+ * The function checks if a given string is a valid JSON object.
+ * @param {string} value - The parameter `value` is a string that represents a JSON object.
+ * @returns The function isJSONObject returns a boolean value. It returns true if the input value is a
+ * valid JSON object, and false otherwise.
+ */
+export default function isJSONObject(value: string): boolean {
+  if (typeof value !== 'string') {
+    return false;
+  }
+  try {
+    const parsedValue = JSON.parse(value);
+    return typeof parsedValue === 'object';
+  } catch (error) {
+    return false;
+  }
+}
