+++ src/parseToBoolean.ts
@@ -1,0 +1,14 @@
+/**
+ * The function `parseToBoolean` takes a string as input and returns a boolean value if the string is
+ * either 'true' or 'false', otherwise it logs an error message and returns undefined.
+ * @param {string} val - The `val` parameter is a string that represents a boolean value.
+ * @returns The function `parseToBoolean` returns a boolean value if the input string can be parsed to
+ * either `true` or `false`. If the input string is not equal to either `true` or `false`, it logs an
+ * error message to the console and returns `undefined`.
+ */
+export default function parseToBoolean(val: string): boolean | undefined {
+  if (val === 'true') return true;
+  if (val === 'false') return false;
+  console.error('String cannot be parsed to boolean.', val);
+  return undefined;
+}
